requirements.txt
django
djangorestframework
drf-extensions

#db
mysql-connector-python

#schemas
coreapi
pyyaml